Move parseCloudfuseError into engine.cpp as a static helper function
